% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_qp_f.R
\name{box_qp_f}
\alias{box_qp_f}
\title{Quadratic programming with box constrains}
\usage{
box_qp_f(Q, u, b, rho, Maxiter = 10^3, tol = 10^-4)
}
\arguments{
\item{Q}{symmetric p x p  matrix}

\item{u}{optimization variable, a vector of length p}

\item{b}{is a vector of length p, this is a problem parameter.}

\item{rho}{tuning parameter}

\item{Maxiter}{maximum number of itarations; default is 10^3}

\item{tol}{convergence torlerance; default is 10^-4}
}
\description{
This function is borrowed from dpglasso package
}
