% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVsglasso.R
\name{CVsglasso}
\alias{CVsglasso}
\title{Cross-Validation for Simplified Graphical Lasso.}
\usage{
CVsglasso(
  X,
  lambda = 10^seq(-2, 2, 0.2),
  niter = 30,
  tolerance = 1e-04,
  thresh = 1e-04,
  pen.diag = FALSE,
  path = FALSE,
  adjmaxit = NULL,
  K = 5,
  crit.cv = c("loglik", "AIC", "BIC"),
  start = c("warm", "cold"),
  cores = 1,
  trace = c("progress", "print", "none"),
  ...
)
}
\arguments{
\item{X}{n x p data matrix}

\item{lambda}{list of tuning parameters, default is \code{10^seq(-2, 2, 0.2)}}

\item{niter}{maximum number of outer itarations; default is 30}

\item{tolerance}{tolerance value for convergence; default is "ave" which the relative difference in the Fronenius norm of the precision matrices across two successive itarations}

\item{thresh}{threshold value}

\item{pen.diag}{whether to penalize diagonal elements; \code{TRUE} by default}

\item{K}{numer of fold; default is 5}

\item{crit.cv}{CV criterian; default is \code{loglik}}

\item{start}{algorithm starting criterion; warm by default}

\item{...}{other \code{sglasso()} parameters}
}
\value{
\code{theta} p x p estimated precision matrix

\code{niter} number of iterations until convergence
}
\description{
The main funcitonality is borrowed from \code{CVglasso} package.
}
