% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GenAtHum}
\alias{GenAtHum}
\title{Gene Atlas Human (GenAtHum) Data}
\format{
A list of length 6 containing:
\describe{
  \item{X}{Design matrix of dimension 158 x 2045.}
  \item{y}{Response vector of length 158.}
  \item{group}{Group membership code for each predictor.}
  \item{gene_code}{Vector of gene codes.}
  \item{gene_name}{Vector of gene names.}
  \item{groups_name}{Vector of group names.}
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/}
}
\description{
The microarray data were obtained from the GDS DataSet of the Gene
Expression Omnibus (GEO) repository in the NCBI archives.
}
\details{
This data set contains 158 samples, 2045 predictors, and 79 groups,
as described in Yuzbasi and Cao (2025).
}
\examples{
data(GenAtHum)
X <- GenAtHum$X
y <- GenAtHum$y
group <- GenAtHum$group

}
