% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.R
\name{logLik.sglasso}
\alias{logLik.sglasso}
\title{Log-likelihood for sglasso}
\usage{
\method{logLik}{sglasso}(object, ...)
}
\arguments{
\item{object}{A fitted \code{sglasso} object.}

\item{...}{Additional arguments.}
}
\value{
Log-likelihood value.
}
\description{
Log-likelihood for sglasso
}
