\name{neisseria}
\docType{data}
\alias{neisseria}
\title{
Neisseria Data Set
}
\description{
This data set contains the gene expression data from a high-resolution time-course experiment besed on the sequenced Neisseria meningitidis serogroup strain B strain MC58. Specifically, the expression level of ten genes is measured at ten different time points. Each column is standardized to have zero mean and standard deviation equal to one.
}
\usage{data("neisseria")}
\keyword{datasets}
