% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select.sglasso}
\alias{select.sglasso}
\title{Select tuning parameters for sglasso}
\usage{
\method{select}{sglasso}(obj, criterion = c("AIC", "BIC", "EBIC"), ...)
}
\arguments{
\item{obj}{A fitted \code{sglasso} object.}

\item{criterion}{Selection criterion.}

\item{...}{Additional arguments.}
}
\description{
Select tuning parameters for sglasso
}
