\name{sglasso-internal}
\title{Internal sglasso functions}
\alias{sglasso.fit}
\alias{make_sglasso}
\alias{make_action}
\alias{setDiff}
\alias{mask2Tv}
\alias{mask2Te}
\alias{theta2Kh}
\alias{fglasso_model2mask}
\alias{zero}
\description{Internal sglasso functions}
\author{Luigi Augugliaro}
\details{These are not intended for use by users}
\keyword{internal}
