% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sglasso-package.R
\docType{package}
\name{sglasso-package}
\alias{sglasso-package}
\title{sglasso: Scaled Group Lasso}
\description{
Tools for scaled group lasso estimation and cross-validation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/byuzbasi/sglasso}
  \item Report bugs at \url{https://github.com/byuzbasi/sglasso/issues}
}

}
\keyword{internal}
