% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.sglasso.R
\name{coef.sglasso}
\alias{coef.sglasso}
\title{Extract coefficients from a sglasso object}
\usage{
\method{coef}{sglasso}(object, lambda, d, which = 1:length(object$lambda), drop = TRUE)
}
\description{
Extract coefficients from a sglasso object
}
