% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.sglasso.R
\name{cv.sglasso}
\alias{cv.sglasso}
\title{Extract coefficients from a cv.sglasso object.}
\usage{
cv.sglasso(
  X,
  Y,
  group = 1:ncol(X),
  lambda,
  nlambda = 100,
  d,
  nd = 11,
  alpha = 0.5,
  nfolds = 10,
  fold,
  beta_start = NULL,
  family = "gaussian",
  bilevel = F,
  max_iter = 1e+08,
  eps = 1e-04
)
}
\arguments{
\item{object}{Fitted cv.sglasso object.}

\item{...}{Additional arguments for compatibility.}
}
\value{
An object of class cv.sglasso, a list with entries
}
\description{
Extract coefficients from a cv.sglasso object.
}
\examples{
data(Birthwt,package = "grpreg")
X <- Birthwt$X
group <- Birthwt$group
Y <- Birthwt$bwt
CVsglasso_fit <- cv.sglasso(X,Y,group)
}
\seealso{
\code{\link{sglasso}}
}
