% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GenAtHum}
\alias{GenAtHum}
\title{Gene Atlas Human (GenAtHum) Data}
\format{
An object of class list of length 6.
}
\usage{
GenAtHum
}
\arguments{
\item{X}{is a design matrix [158 x 2045]}

\item{y}{is a response vector [1 X 158]}

\item{group}{is the code for each group name}

\item{gene_code}{is the vector of code for each gene}

\item{gene_name}{is the vector of names for each gene}

\item{groups_name}{is the vector of names for each groups}
}
\description{
The microarray data is obtained from the GDS DataSet of the Gene Expression Omnibus (GEO) repository in the NCBI archives (\url{www.ncbi.nlm.nih.gov/geo})
}
\details{
This data set contains 158 samples with 2045 predictors, and 79 groups as described in Yuzbasi and Cao (2025).
}
\examples{
data(GenAtHum)
X <- GenAtHum$X
y <- GenAtHum$y
group <- GenAtHum$group
}
\keyword{datasets}
