% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.sglasso.R
\name{predict.sglasso}
\alias{predict.sglasso}
\title{Predict method for sglasso objects}
\usage{
\method{predict}{sglasso}(
  object,
  newx,
  type = c("response", "coefficients", "vars", "groups"),
  lambda,
  d,
  s = "ALL",
  opt_beta = NULL,
  which = 1:length(object$lambda),
  drop = TRUE,
  ...
)
}
\arguments{
\item{object}{Fitted sglasso object.}

\item{newx}{New design matrix.}

\item{type}{Prediction type.}

\item{lambda}{Lambda values.}

\item{d}{Scaling parameter values.}

\item{s}{Selection mode.}

\item{opt_beta}{Optional coefficient vector.}

\item{which}{Indices of lambda values.}

\item{drop}{Logical; should dimensions be dropped?}

\item{...}{Additional arguments.}
}
\description{
Predict method for sglasso objects
}
