% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select.sglasso}
\alias{select.sglasso}
\title{Select tuning parameters for sglasso}
\usage{
select.sglasso(
  obj,
  criterion = c("AIC", "BIC", "EBIC", "GCV", "AICc"),
  ebic_gamma = 0.5,
  ebic_level = c("group", "feature"),
  tol = 1e-08,
  ...
)
}
\arguments{
\item{obj}{A fitted \code{sglasso} object.}

\item{criterion}{Selection criterion.}

\item{ebic_gamma}{EBIC gamma parameter. Default is 0.5.}

\item{ebic_level}{EBIC penalty level. Use \code{"group"} for group-level
EBIC or \code{"feature"} for feature-level EBIC.}

\item{tol}{Tolerance used to determine nonzero coefficients.}

\item{...}{Additional arguments.}
}
\description{
Select tuning parameters for sglasso
}
