% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select.sglasso}
\alias{select.sglasso}
\title{Extract coefficients from a sglasso object}
\usage{
\method{select}{sglasso}(obj, criterion = c("BIC", "AIC", "GCV", "AICc", "EBIC"))
}
\description{
Extract coefficients from a sglasso object
}
