% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sglasso.R
\name{sglasso}
\alias{sglasso}
\title{Simplified Graphical Lasso}
\usage{
sglasso(
  S,
  init_theta = NULL,
  lambda = 0.1,
  pen.diag = TRUE,
  niter = 30,
  inner_niter = 1000,
  tolerance = 1e-05,
  inner.tolerance = 1e-07,
  thresh = 1e-04,
  tol.type = c("max", "ave", "F"),
  trace.objfnct = FALSE
)
}
\arguments{
\item{init_theta}{p x p initial positive definite precision matrix}

\item{lambda}{tuning parameter, default is 0.1}

\item{pen.diag}{whether to penalize diagonal elements; \code{TRUE} by default}

\item{niter}{maximum number of outer itarations; default is 30}

\item{inner_niter}{maximum number of inner iterations; default is 1000}

\item{tolerance}{tolerance value for convergence;}

\item{inner.tolerance}{tolerance value for convergence of the inner loop;}

\item{thresh}{threshold value}

\item{tol.type}{default is "max" which the relative difference of the maximum of precision matrices across two successive iterations}

\item{trace.objfnct}{whether to compute objective function. Increases computation time}

\item{X}{n x p data matrix}
}
\value{
\code{theta} p x p estimated precision matrix

\code{niter} number of iterations until convergence
}
\description{
This function implements GLasso using modifications proposed in
Dallakyan and Pourahmadi (2024)
}
