% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sglasso.R
\name{sglasso}
\alias{sglasso}
\title{Extract coefficients from a cv. sglasso object.}
\usage{
sglasso(
  X,
  Y,
  group = 1:ncol(X),
  lambda,
  nlambda = 50,
  d,
  nd = 11,
  alpha = 0.5,
  beta_start = NULL,
  family = "gaussian",
  bilevel = F,
  max_iter = 1e+08,
  eps = 1e-04,
  lambda.min = 0.005,
  dfmax = p,
  gmax = length(unique(group))
)
}
\arguments{
\item{object}{Fitted sglasso object.}

\item{...}{Additional arguments for compatibility.}
}
\value{
A vector of coefficients
}
\description{
Extract coefficients from a sglasso object.
}
\examples{
data(Birthwt,package = "grpreg")
X <- Birthwt$X
group <- Birthwt$group
Y <- Birthwt$bwt
sglasso(X,Y,group)
}
\seealso{
\code{\link{sglasso}}
}
