% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.sglasso.R
\name{transform_sglasso_betas}
\alias{transform_sglasso_betas}
\title{Extract coefficients from a fitted sglasso object}
\usage{
transform_sglasso_betas(object)
}
\arguments{
\item{object}{A fitted \code{sglasso} object.}

\item{lambda}{Lambda values for coefficient extraction.}

\item{d}{Scaling parameter values.}

\item{which}{Indices of lambda values.}

\item{drop}{Logical; should singleton dimensions be dropped?}

\item{...}{Additional arguments for compatibility.}
}
\value{
Estimated coefficient array or matrix.
}
\description{
Extract coefficients from a fitted sglasso object
}
